+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Fixes
 
 - Do not include `BuildConfig` into source bundles ([#705](https://github.com/getsentry/sentry-android-gradle-plugin/pull/705))
+- Fix misleading auth-token error message in case "sentry-cli info" fails ([#708](https://github.com/getsentry/sentry-android-gradle-plugin/pull/708))
 
 ### Dependencies